+++ internal/wordmap/markdown_test.go
@@ -1,4 +1,4 @@
-package dicts
+package wordmap
 
 import "strings"
 import "testing"
@@ -119,6 +119,35 @@ func TestMarkDownTwoTables(t *testing.T) {
   if actual != expected {
     t.Errorf("Incorrect to-value at index 1 (actual %s)", actual)
   }
+}
+
+func TestMarkDownEmptyColumnValues(t *testing.T) {
+  t.Run("Empty from value", func(t *testing.T) {
+    markdown := `
+      | from | to  |
+      | ---- | --- |
+      |      | bar |
+    `
+
+    _, err := parseMarkDownFile(&markdown)
+
+    if err == nil {
+      t.Errorf("Error should be set if the from value is empty")
+    }
+  })
+  t.Run("Empty to value", func(t *testing.T) {
+    markdown := `
+      | from | to |
+      | ---- | -- |
+      | foo  |    |
+    `
+
+    _, err := parseMarkDownFile(&markdown)
+
+    if err == nil {
+      t.Errorf("Error should be set if the to value is empty")
+    }
+  })
 }
 
 func TestMarkDownIncorrectHeader(t *testing.T) {